Update patch set 1
Patch Set 1:

Hey, sorry for the hiccup!

> We have a new binary "vbmcd". Does it require its own systemd unit file?

There is the new executable *vbmcd*, it is the master process which now manages all the per-domain virtualbmc instances.

The way how this change is done is that if you run *vbmc* in the old way and *vbmcd* is not running on the system, *vbmc* will invoke *vbmcd* behind the scenes (this is a compatibility feature). In that sense you could potentially ignore *vbmcd*.

But! The whole point of this change is to run *vbmcd* from systemd what gives us way more control, convenience and reliability. Therefore it would be much better to introduce systemd unit for *vbmcd*!

Does it make sense?

Patch-set: 1
Label: Code-Review=0